// src/enums/callback-command.ts
export enum CALLBACK_COMMAND {
  showState = 'state',
  listCategories = 'lstCat',
  createCategory = 'creCat',
  selectCategory = 'selCat',
  createCategoryCancel = 'creCatCan',
  editCategory = 'edCat',
  editCategoryCancel = 'edCatCan',
  deleteCategory = 'delCat',
  deleteCategoryCancel = 'delCatCan',
  deleteCategoryConfirm = 'delCatCon',
  addCard = 'addCar',
  addCardCancel = 'addCarCan',
  addCardEmptyAnswer = 'addCarEmp',
  editCardQuestion = 'edCarQ',
  editCardQuestionCancel = 'edCarQCan',
  editCardAnswer = 'editCarA',
  deleteCard = 'delCar',
  deleteCardCancel = 'delCarCan',
  deleteCardConfirm = 'delCarCon',
  showNextCard = 'nextCar',
  skipCard = 'skipCar',
  showAnswer = 'showA',
  markAnswer = 'markA',
}